DRT-4607 Refactor
